apps/collector/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "collector"
version = "0.1.0"
description = "Funding & company data fetcher"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.26.0",
    "pyyaml>=6.0",
    "pandas>=1.3.0",
    "tqdm>=4.62.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.5b2",
    "isort>=5.9.0",
    "mypy>=0.910",
]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 